src/teams_integration/teams_bot.py
# Microsoft Teams Integration for Master Item AI Agent

from botbuilder.core import BotFrameworkAdapter, TurnContext
from botbuilder.schema import Activity

class MasterItemBot:
    def __init__(self):
        pass

    async def on_message_activity(self, turn_context: TurnContext):
        """
        Handle incoming messages from Microsoft Teams.
        """
        user_message = turn_context.activity.text
        response = self.process_message(user_message)
        await turn_context.send_activity(Activity(type="message", text=response))

    def process_message(self, message):
        """
        Process the user's message and return a response.
        """
        # Example: Simple echo bot
        return f"You said: {message}"

if __name__ == "__main__":
    print("Microsoft Teams bot is ready to be integrated.")
